Move clipServiceUtils.js to popup directory and update import path
